Added Bank

+++ Person.cc
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-Person::Person(std::string n, int e)
+Person::Person(std::string n, float e)
 :name{n}, expence{e}, debt{0}
 //  ----------------------
 {}
@@ -12,25 +12,25 @@ string Person::get_name() const
     return name;
 }
 
-void Person::update_debt(int cost)
+void Person::update_debt(float cost)
 //  ----------------------
 {
     debt += cost;
 }
 
-int Person::get_debt() const
+float Person::get_debt() const
 //  ----------------------
 {
     return debt;
 }
 
-int Person::get_expence() const
+float Person::get_expence() const
 //  ----------------------
 {
     return expence;
 }
 
-void Person::update_expence(int cost)
+void Person::update_expence(float cost)
 //  ----------------------
 {
 expence += cost;
